refactor: Move result calculation logic to presenter
